requirements.txt
aio-pika==6.8.0
aiormq==3.3.1
idna==3.2
multidict==5.1.0
pamqp==2.3.0
PyYAML==5.4.1
yarl==1.6.3
